test: route Petri usage coverage through provider edge
